backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0